Final Push of code
This is the last update to the code I've done over the summer. The program works and does what is needed, but the results just don't make sense yet. 80% of the work takes up 20% of the time, while the last 20% of the work takes up 80% of the time. Good luck!